requirements.txt
cherrypy==18.1.1
pyyaml
psutil
natasha
googletrans==4.0.0-rc1
ftfy
regex
requests
onnx
onnxruntime
opencv-contrib-python>=4.2.0.34
opencv-python
matplotlib==3.2.1
tqdm==4.45.0
numba==0.48
click
typer
pyspng
ninja
imageio
imageio-ffmpeg==0.4.3
flask==1.1.2
flask_cors==3.0.9
flask_restful_swagger_2==0.35
dlib==19.22.1
tensorboard
lpips